Add `inverse_of` to order.payments
This doesn't seem to happen automatically and is more important now
that we're using cached association objects in OrderUpdater (see
Solidus PR 1389).  Solidus 1.4 broke some of our specs without it.

In a Solidus sandbox without this patch:

    >> o = Spree::Order.create!
    >> p = o.payments.create!
    >> puts p.order.object_id, o.object_id
    70240633410460
    70240634674300

With this patch:

    >> o = Spree::Order.create!
    >> p = o.payments.create!
    >> puts p.order.object_id, o.object_id
    70145793491740
    70145793491740

There is some important history around this particular `inverse_of` --
see solidusio/solidus PR 407 and spree/spree_gateway PR 132.

Commit d7b8c73 in Solidus removed part of the specs around this,
assuming that the `inverse_of` was automatically added, but it doesn't
actually seem to be (see above). The remaining tests around the issue
still pass.