Add "only_allow_most_recent_promotion" configuration
Add an `only_allow_most_recent_promotion` config that allows stores to
only allow a single promotion per order and for that promotion to be
the promotion most recently applied by the user.

This is a somewhat common way to work with promotions and helps reduce
unintended side effects from promotions interacting with each other.
E.g. accidentally giving more of a discount than a store intends to
when a clever customer applies multiple promotions.

This also requires adjusting PromotionHandler::Cart so that it won't
automatically apply "apply_automatically" promotions if this config is
enabled and the order in question already has a promotion applied to
it.