DummyGenerator doesn't need to create a custom_user.rb initializer
The InstallGenerator, which is always going to be called after the
DummyGenerator, includes a `Spree.user_class` setter in
initializers/spree.rb, which is also parameterized to use options[:user_class]
https://github.com/solidusio/solidus/blob/d7b7ef47087e998655481a467ec3bf6d9a658ce3/core/lib/generators/spree/install/templates/config/initializers/spree.rb#L79

There's no reason for DummyGenerator to create a hard-coded commented-out
Spree.user_class setter in a separate initializer. It's dead code, but
makes things somewhat more confusing for someone trying to decipher
the generator(s).