Convert blank country_iso on prices to nil
When adding a new price for a product in the admin panel using the
prices tab and not selecting a country for this price ("Any Country"),
the browser will send an empty string as country iso. In the backend,
the intention of this user action is for the price to be saved with a
`country_iso` of `nil`, indicating that it is a fallback price valid
for any country.

This commit converts an empty string `country_iso` to `nil` before
validation, allowing admin users to actually create a new price.

Fixes https://github.com/solidusio/solidus/issues/1453