Add ledger class to to keep track of financial history for store credits
The `store_credit_event` is recording an audit trail to provide some
insight in the lifeline of a `store_credit`. However, it would be nice to
have an actual ledger for the store credits, that will focus only on the
events that will have a financial consequence to the store owner.

This PR introduces the `store_credit_legder_entry` that will provide
historically consistent reporting of store credit liabilities and amount
changes.

* Add `StoreCreditLedgerEntry#credit` with class level methods `credit`
* and `debit` that will create the correct entries for a specific
`store_credit`.
* Add `StoreCreditLedgerEntry#balance` that will show the remaining
amount that the `store_credit` user has available.
* Add `StoreCreditLedgerEntry#liability_balance` that will show the
liability amount for the specific `store_credit`.
* Update the admin `store_credit#show` view that will list all the above
amounts and entries.
* Add a migration that will create the correct initial
`StoreCreditLedgerEntry`records for all existing `store_credit` records.
This will create a `StoreCreditLedgerEntry` for both liability amounts
and pending amounts.
* Refactor `Spree::Admin::StoreCreditEventsHelper` and introduce
`Spree::Admin::StoreCreditOriginatorHelper` to better show the links to
the originator classes.
* Create the correct `StoreCreditLedgerEntry` records for all calls made
on the `StoreCredit` that is changing amounts for the `StoreCredit`.
* Add specs for all introduced new functionality and expectations.
* Update CHANGELOG