Add amount_remaining for Spree::StoreCreditEvent
This additional field will make sure that credit events will record the
actual unused amount for a particular Spree::StoreCredit.

This also hides the user_total_amount in the backend because it makes no
sense to display the total credit a user has at her/his disposal when
visiting the store credit show route in the backend.