Changed latest_payment to call reorder.
Issue #1525
Currently latest_payment key in cart.html.erb is set to `@order.payments.order(...)`.
However because payments as a default scope set to `:created_at` The order method will
be ignored and latest_payment will actually be set to the earliest payment.  To get the
latest payment the method `reorder` should be called to override the default scope's order.