Allow using ActiveModel::Naming for calculator descriptions
The descriptions for calculators are not namespaced. This commit makes
`Spree::Calculator.description` are shortcut to `Spree::Calculator.model_name.human`.

This way, we can eliminate a few lines from the codebase, as well as
converging around Rails conventions for naming things.

This commit enables this technique and takes the default tax calculator
as an example.