Correct behavior of Order#outstanding_balance
This commit addresses an issue in `Spree::Order#outstanding_balance`, in
circumstances where an Order has had refunds applied.

The existing behaviour cancels out the amount of refunded payments,
resulting in a negative outstanding balance and an order state of
'credit_due', when in fact the order balance is zero. This is caused by
the calculation of `payment_total` already factoring in refunds.

Some more discussion about this behaviour can be found in the following
issues:

https://github.com/solidusio/solidus/issues/1107
https://github.com/solidusio/solidus/pull/1536

With these changes, `outstanding_balance` is now defined as
`total - payment_total`. This corrects the behaviour around refunds.

Tests have been updated accordingly, with some help from
DanielePalombo's commit:

https://github.com/solidusio/solidus/pull/1536/commits/c851404860a43f07a77d2c9a80dd6afcf2a54984