Avoid re-creating tax adjustments
Previously, ItemAdjuster#adjust! would always destroy all tax
adjustments on the item and recreate them. This was unnecessarily slow
in the common case of needing no changes (especially following #1479).

This changes ItemAdjuster#adjust! to instead update the adjustment for
a rate if it already exists, create adjustments if they are missing for
a matching rate, and destroy adjustments for rates which are no longer
matching.