Simplify image upload UI
This replaces the new image upload area (previously a paragraph of text
and a file upload field) with a single button inside the drop area with
the text "Browse or drag files here to upload".

The original problem I set out to fix was that after uploading an image
through the file input, the filename remained. I added some javascript
to clear the field after uploading, which worked great, but now there
was always the text "No file selected". After this I realized the entire
area was a label and we can just hide the upload field.