Only run setup_for_current_state in edit, update
Because that before_action was running on all actions of the controller,
the extra actions added by solidus_auth_devise had this run before them,
which wasn't desirable.

This has been fixed via a skip_before_action in solidus_auth_devise, but
we might as well restrict it here as well, since it only makes sense for
these two specific actions.

See also
https://github.com/solidusio/solidus_auth_devise/pull/76 and
https://github.com/solidusio/solidus/issues/1588