Add a guide for pricing and taxation
This is a re-do of the Spree Taxation guide, adjusted for the changes in
Solidus 1.3 and up. It explains how to setup your store for VAT-style taxation
as well as Sales tax. Because of how VAT-style taxation influences prices,
this document also explains how the price selector/pricing options mechanism works.