Show term and condition to user when required acceptance
When the preference `require_terms_and_conditions_acceptance` is true, the button
`place_order` is disable and before the button appear a text with
checkbox. The textarea contains the terms and conditions of checkout, and
is required to check the checkbox to remove disabled state from
`place_order` button, and procede to the next state.