Fix in_stock cache clearing after update
There's an issue with the caching on `Spree::Variant#in_stock?` in which
it won't be cleared after the variant is updated. Only after it is
touched.

As a result, changes to `Spree::Variant#track_inventory` do not
correctly update `Spree::Variant#in_stock?` when they should.

The simplest solution is to ensure we clear the cache on save in
addition to touch.