Only remove index if index exists
Our app didn’t have a pre-existing index for `spree_products.slug` (and I didn’t see an earlier migration for that), so it was throwing

```ruby
ArgumentError: Index name 'index_spree_products_on_slug' on table 'spree_products' does not exist
```

when we tried to run this migration. This change will make the migration work correctly whether or not the index already exists.