Don't require address#empty? for use_billing
Previously, this would only check the box if the two addresses were
identical AND they were both "empty".

Instead, this now checks "Use billing" any time the two addresses are
equal and the ship_address is a new record. I think this was probably
the original intention of checking empty? (only new records can be
empty because of Address's validations).