Improve AvailabilityValidator spec
Instead of relying on the behaviour of stock_location_quantities and the
entire order state machine, we should just create shipments and
inventory units in the state we expect.

This adds tests for the line item needing some stock from each stock
location.

This adds one skipped test which fails because the error on the line
item is added twice.