Don’t set default text highlight colors
One of the problems with setting `::selection` colors is that it’s nontrivial for people who actually _want_ to use the browser-default selection colors to override the setting in a way that reverts to the browser default. For this reason I think it would be better to simply leave them out of the default `frontend` style and let individual site owners make their own decisions.