Add fix for removing roles from users
In HTML unchecked checkboxes won’t get submitted in a form. That’s why Rails’ `check_box` helper always sets a hidden field with the same name. This field will be submitted when the checkbox is not checked and will be ignored when the checkbox is set.

As we don’t use Rails’s `check_box` helper but `check_box_tag` helper in the admin users form role checkboxes we need to add this hidden field ourselves.