Fix creating superfluous InventoryUnits
This fixes an issue where OrderInventory would miscount the number of
existing inventory_units on a line item and create extra records.

InventoryUnits exist through the line_item.inventory_units,
shipment.inventory_units, and order.inventory_units associations.
Because of this, their associations can't be kept up to date in memory.

OrderInventory was counting inventory using inventory_units.size, which
would return the in-memory count if the line_items.inventory_units
association was loaded.

This issue was exposed by adding to the same line item mutliple times
without reloading the order. The line_item.inventory_units association
would be become `loaded?`, because ActiveRecord's `#size` marks a record
as loaded if the count returns 0 (TIL!). On the next pass it would see
the wrong size because the association was out of date, and add two
inventory units instead of one.

This avoids the issue by counting inventory using inventory_units.count,
which doesn't consider the in-memory association.