Convert admin orders table into full width layout
The admin orders table is too narrow on modern screen sizes. There is plenty of space we could use to display the full information on the orders table.

When adapting the full width layout we get too wide columns on the search form, though. This is fixed by utilizing different column sizes for the viewports. The result is pretty good without moving around any html elements (what potentially could break admin overrides of extensions or shops)