Do not run callbacks when deleting dependent inventory units
An inventory unit should not exist without its line item and its shipment.
The protective callbacks on `Spree::InventoryUnit` just make sure that
one can not destroy an inventory unit for a shipment that cannot be destroyed.

Added a safeguard so that we're not deleting line items with shipped inventory units.