Raise font size in fieldsets.
All fieldsets (and we use lots of them) have a font size of 90% (11.3px). With our base font size of 13px this leads to very small font sizes (10.53px) in all forms inputs (that have an additional 90% font-size) as most of them sit in a fieldset.

Also removes a lot of styles already defined by Bootstrap and styles that needed to be resetetted as the fieldset was defined as part of the default input field styles.