Fix complete order factory to have non-pending inventory units
When an order is completed, all its inventory units are by default
set to `pending: false`. The factory should reflect that.

This made a change to the shipment spec necessary. That spec is setup with
a completed order, and then runs `Spree::Shipment#finalize`, which is
run at order completion, so I manually reset the inventory unit to pending.