Don't display inactive payment methods on the frontend
We use `Spree::Order#available_payment_methods` to determine which
payment methods should be displayed on the payment page in checkout. The
changes in #1540 accidentally dropped the scoping for active payment
methods, causing inactive payment methods to show up on the frontend.

This also required fixing the `available_to_store` scope, which behaved
incorrectly. Calling `self` from inside the scope returns the
`Spree::PaymentMethod` class itself and not the current scoping as the
code implied. This scope will now raise an error if the store parameter
is `nil`, return all payment methods in the current scope if the store
has no payment methods, and return the payment methods for the store if
it does have some.