Use `order` instead of `sort_by` on AR relation
The original implementation of `Spree::PaymentMethod.available` used
`.select` which would turn the ActiveRecord::Relation into an array and
meant we had to use `sort_by`.

In #1540 `available` was deprecated in favour of `available_to_users`
and `available_to_admin`, each of which returns an
ActiveRecord::Relation object. That means we can use `order` to do the
sorting in SQL, which also allows users to chain more scopes onto the
return value if they wish.