Allow promotion list to be sortable.
Update promotion list to have same sort funtionality as most of the rest
of the app. Promotion uses is not sortable as it is a calculated value. We
could actually make it sortable with a custom ransacker. Something like:

    ransacker :usage_count do
      sql = Spree::Adjustment.eligible.promotion.
        where('source_id = spree_promotion_actions.id').
        joins(:order).merge(Spree::Order.complete).
        select('COUNT(order_id)').to_sql
      Arel.sql "(#{sql})"
    end

But I think the complexity of adding that is not worth the benefit of being
able to sort (and search) by usage count.