Remove `distinct` filtering.
It's not clear why this is here. Usually this is used when a child table
is joined to the parent table producing many rows with the same data. There
is no joins here so it doesn't seem necessary.

In addition to being unnecessary it actually causes problems. There is a
hook here were applications can extend the relationships preloaded. But
if they do that they cannot add those to the sort as the DISTINCT will
require all values that appear in the `ORDER BY` to also appear in the
`SELECT`. But removing the `DISTINCT` we give applications more flexibility
in customization.