Set belongs_to_required_by_default = false
As of Rails 5, belongs_to is now required by default. Our models,
including those in extensions and applications, aren't ready for this.

In Rails 5.0 there seems to be some bug that was preventing this new
default from affecting Spree::Base models.

    > Spree::Base.belongs_to_required_by_default
    => nil
    > config.active_record.belongs_to_required_by_default
    => true

In Rails 5.1 this seems to be fixed so we need to be explicit about what
we use. We might as well be explicit in Rails 5.0 anyways.