Remove code sorting.
For stores that are using a LARGE number of codes (i.e. millions) this
setup won't work as it will load all codes for each promotion displayed.

Also under a multi-code setup, sorting doesn't make much sense.

Due to these problems I am backing the sorting of this column back out
and just going to make that a local change (since the store this is for
is generally just uses one code per promotion).