Fix styling of tiered promotions delete icon
Previously this trash icon was being partially obscured by the row next
to it, and only the left half of the icon could be clicked. It was also
floating outside of the parent container. This commit fixes these issues
by adding a margin to the row, and placing the icon inside.

It also changes the cursor to a pointer on hover.