Remove click event from payments diplay amount
In order payments table the amounts are currently clickable.
As the edit icon is only displayed if the payment is pending,
the amount is always clickable.

Instead of making this conditional I removed this feature all together
as I don't see any of this behaviour in any other admin table.