Show focused border on currency selector
Previously we were embedding the currency selector inside of a bootstrap
input-group-addon and hiding its border. This required a little extra
magic because we had to hide the select's border and change its height
to fit within.

Having no border meant that there was no indication when the currency
input was focused.

Bootstrap doesn't fully support having multiple inputs within a
input-group, but we can make it work by hiding the right border on the
left input.

A slight hack is needed to make focus look correct: we show the focused
colour on the right input's left border when the left border is focused.
This achieves the desired effect.