Remove fallback first shipping method on shipments
In the case where the system has not picked a shipping rate, or
the user has not picked one, we can not just fall back on the first
shipping rate. In this case, the "selected" shipping rate's amount
and the shipment's amount will be out of sync, which is inconsistent.

This commit removes the fallback.