Fix sidebar alignment
The sidebar is currently not correctly vertically aligned to the fieldset it most of the time is displayed with (ie. in the order detail views and the user detail views).

Use a relatively unit for top margin instead and move the top border to the heading to respect the column gutter.