Allow refreshing rates for unshipped complete orders
When an order is complete, but not shipped yet, admins need to be able
to refresh rates for a shipment.

This did not work for two reasons:
- `#refresh_rates` was not called on a new shipment
- If it had been called, it would not have done anything.

I added an expectation for a new shipment to be generated with a selected
shipping rate to test this behaviour.

This is a less scary change than it seems since the order updater does not
refresh rates automagically anymore.