Deprecate simple_current_order
The `#simple_current_order` method is duplicating many of the
functions already found in `#current_order`. `simple_current_order`
was finds an order by token or user, updates ip_address, and
creates an order if one is not found. All functions that are
already accomplised in `#current_order` which also does more
than `#simple_current_order` making it the preferred method.