Add validity period for Spree::TaxRate
This code adds a starts_at and expires_at datetime field to
Spree::TaxRate so that tax rates can be scoped in time.

Null values will be treated as valid so that a tax rate is valid by
default and you can just ignore this feature if you don't need it.

The Spree::Calculator::DefaultTax will always return 0 in case the tax
rate is not valid (either not started or expired).

Ref. #1857