Add handler to create weighted adjustments
We'll be introducing a new calculator for promotions that will allow
order-level promotion values to be applied through line item
adjustments.

This class keeps some of the more complicated logic out of the
calculator as it has to reach up to the order and calculate values on
multiple line items.