Add DistributedAmount calculator for line item adjustments
This calculator allows us to use order level promotions ("Save $50 on
your order") and spread that $50 over all of the line items. This
calculator shoots logic off to another class to handle the calculations.

Each line item needs to call out to the DistributedAmountsHandler
separately which is annoying but each line item needs to be aware of the
order as a whole.

The calculator spreads the discount amount among each of the line items
relative to their price. This means that more expensive items will
receive a greater share of the discount.