Improve order factories
* Add build support for order_with_line_item factory

Changing the `after` callback to use `build` instead of `create` results
in the same behavior in the create case and adds support for the build
case.

* Add test coverage for building order via factory

It's important we establish that factories work correctly when built or
created.

* Improve order_with_totals factory

This factory did not work for the `build` use case. Additionally, it
required a reload.

* Add test coverage for order_with_line_items factory

This test coverage ensures all transient arguments to the factory are
honored as expected.

* Test all child factories of order_with_line_items

The call to `refresh_shipping_rates` in completed_order_with_totals
caused attributes to the factory to be lost.

* Nest completed_order_with_promotion factory

This factory was dangling out of the nesting convention the rest of our order
factories follow.

* Update backend order details admin spec

Now that shipping methods aren't being clobbered by the factory, the
shipping method assignment is working normally.