Uniform payment methods params key
Currently the admin payment methods form uses a mix of payment_method and the payment methods class name params key. This is due to the STI nature of payment methods.

By using the :as feature from Rails' form_for helper we remove code from the controller that merges these to kind of params hashes into one and reduce clutter in the form.