Rename PaymentMethod#provider to gateway
In order to remove confusing parts of our current payment methods implementation we rename the `provider` methods from `Spree::PaymentMethod` into `gateway`.

A gateway is a class that implements the API communication to the provider. See the `active_merchant` gem for a set of gateway classes we use in `solidus_gateway`.

A provider on the other hand is a service vendor that offers a payment gateway and multiple payment methods (like Braintree).

Therefore we plan to introduce a provider class later on. This then holds commonalities for mulitple payment methods, like credentials. Also it will be an ActiveRecord model so that payment methods can be grouped under one provider, which is very useful for BI and reporting.