Dynamically define methods for payment sources
(This is a work in progress)
In the wallet define methods for each descendant of the `Spree::PaymentSource` class.
This will allow users to do the following:

```ruby
user.wallet.credit_cards
user.wallet.gift_cards
```

The reason this is a work in progress is that this will not work
expectedly in Development. Because of lazy loading, the descendants may
not be loaded yet causing `Spree::PaymentSource.descendants` to return
an empty array.  Due to this, it may be best to not do this dynamically.