Make _report_order_criteria more shareable
When creating a custom report on the reports controller, I noticed
that this shared form would always post the the :sales_total action of
the reports controller (`Spree::Admin::ReportsController`).

See
https://github.com/solidusio/solidus/blob/v2.1/backend/app/views/spree/admin/shared/_report_order_criteria.html.erb#L1

In order to allow other reports be able to actually share this partial,
pass the key into the partial so that the partial will post to the right
action.