Select all shipping rates in the Estimator
This change selects all shipping rates using the Estimator and
dictates the selection of those rates to the various views. For
example, in the backend views we don't limit what rates are shown
whereas on the frontend we limit to only what users can see or
if a backend rate is already selected display that.