Rewrite Product.available using WHERE EXISTS
This avoids returning duplicate products from Product.available.

Previously a product would be returned multiple times from the
Product.available scope if it had multiple prices on its master variant.

This wasn't an issue when products were returned from a search or on the
default frontend because .distinct was used, but it was confusing when
.available was used directly.

This replaces the INNER JOIN to prices with a WHERE EXISTS subquery.