Emit Spree.url deprecation warning in all envs
This has been deprecated forever, but I'm still not confident enough to
remove it. It's possible that it hasn't been seen by users because
Spree.env has been unset (as it is in our default frontend).

This commit changes the warning to be issued always (so long as console
and console.warn are defined).