Deprecate Shipment#add_shipping_method
This was a convenience method which was used in specs to add a shipping
rate with a given shipping method. However it isn't suitable for use
outside of specs it doesn't handle having an existing selected rate and
doesn't calculate the correct cost.

In specs it would be better just to create the shipping_rate instead of
using this method, as it will be more obvious what is going on.