Remove ! from assign_default_user_addresses!
There's no version of this method without an exclamation mark.
Additionally, this doesn't have any of the normal tells of a ! method,
it doesn't make changes to the database, shouldn't raise exceptions, and
doesn't do anything "dangerous".

This commit renames the method to assign_default_user_addresses, and
adds assign_default_user_addresses! as a deprecated alias.