Fix admin SQL issues with DISTINCT products
MySQL 5.7+ previously errored when trying to sort products by price.

ORDER BY clause is not in SELECT list,
references column 'spree_prices.amount' which is not in SELECT list;
this is incompatible with DISTINCT

Additionally, under PostgreSQL, searching for an SKU which matches
multiple variants of the same product and then sorting by price,
multiple of the same Product would be returned in the search.

This commit removes the Product.distinct_by_product_ids hack replaces
the join search on products->variants with a WHERE EXISTS subquery.