Use normal <select> in admin Customer Details step
Previously the country and state input used select2. In the past, using
select2 was the only way to style a select consistently with the rest of
our inputs.

This replaces the select2 on country and state with a normal select
with has bootstrap's custom-select styling applied.

This doesn't replace the "Choose a customer" select2 input, which is
still needed since it searches and fetches ajax data.