Re-add ability to capture partial amounts
Previously when an amount was captured that was smaller than a payment's
total amount, the payment's amount was reduced to the captured sum. Also
the payment was always sent to the `completed` state, regardless of the
captured amount.
Now the payment only transitions to `completed` when the complete amount
is captured (the sum of the capture events is equal to the amount of the
payment). This allows to have multiple capture events per payment without
cloning it, because only a non-complete payment can be captured.