Simplify current store selection
Previously we selected the current store by pretty hacky means. We took
the value of either the SPREE_STORE HTTP header or the domain name, and
then selected a store either matching that code or had that value
contained anywhere in its url attribute. Failing to find these it falls
back to the default store.

Using SPREE_STORE and finding by code is intended to be used with a
reverse proxy like nginx, and is a good and powerful way to configure
Solidus' multi domain. But it does require extra and very specific
configuration, and isn't overly suitable as a default.

Selecting by url which contains the (header or domain name) value is
slow, surprising, and error-prone.

This commit removes support for the SPREE_STORE header, and allows only
one url per-store. Users needing this functionality can replicate it by
implementing their own store selector class.

This commit also allows the store selector to perform only a single
query to find the current store.