Fix flaky spec in stock_management_spec.rb
The same product is already created previously in this spec.

The spec actually expects *not* to use the product that was being created here
-- the sku won't match the `expect` because _that_ sku comes from the first
product created.

This spec fails if the DB happens to return products in reverse-id order in the
admin (which can happen since both products have the same name and the default
sort is by name).