Fix specs for alerts on shipments page
Previously, if a shipment split failed, the page would issue an
alert(undefined). It would never alert with a proper error.

This failed on the capybara selenium driver because it was confused
about getting undefined.

This replaces
     alert(undefined)
with
    alert("Unable to complete split")

Which is a little better (admittedly not much) for users, but doesn't
confuse the selenium test driver.