Avoid running validations in current_order
Previously we updated the IP address on each request by calling

    update(last_ip_address: ip_address)

Though this doesn't run a SQL UPDATE if not necessary, it will still run
validations (resulting in two SELECTs).

This commit adds a new record_ip_address(ip_address) method to
Spree::Order, which will only update the ip_address if it has changed.

This also switches `update` to `update_attributes!`, so that any errors
will be raised.