Add FulfillmentChanger
This class takes two shipments and moves a certain amount of stuff
between them. This used to be called "splitting" a shipment, but really we
are changing fulfillment for those variants, hence the name.

The class has the following features:

1. Speed

It only does four or five requests to the database, instead on one
for every inventory unit moved.

2. Refresh rates after changing fulfilment

If we change how things are fulfilled, the costs for
those fulfilments will change. This change reflects that.

3. Make sure that backordered items are moved first

When changing the fulfilment of an item, we want backordered items
to be changed to a new shipment first. There, we have a chance of them
becoming on_hand (shipped faster).

Because "backordered" is earlier in the alphabet than "on_hand", we can
use an `order` clause in the update SQL statement to achieve this.