Use FulfilmentChanger for transfer_to_* endpoints
With the new FulfilmentChanger, we can stop using
Shipment#tranfer_to_*.

This comes with a few minor API changes: The transfer_to_*
endpoints now respond with a 202 Accepted instead of
201 Created. They only return 4xx responses if the shipment
can not be found or the user is unauthorized.

When changing fulfilment for a shipment, the backend now
generates readable errors and will not return a 4xx JSON response
unless something goes badly wrong. This commit changes the
response handling so our JS code knows whether a split has
been successful, and displays a meaningful error message to
the user.