Fix Order Details spec
This spec had an implicit expectation of all inventory units
of the first shipment to be setup as on hand, but somehow managed to
get into that state by what I think is whacky magic.

Updating the two items to be on hand at the beginning of the spec
makes sure that the remaining item after the first split is also on hand.