Deprecate Spree::OrderCapturing class
This task was added to automate capturing pending payments on a regular basis via a rake task. This kind of custom business logic should not be maintained by us. Half of the class features is not even used by the provided rake task and it makes further improvements around payments harder to test.