Check if deprecated method_type if overridden
method_type has been deprecated in favour of the more descriptive
partial_name. There is a deprecation notice when method_type is called,
but there isn't when method_type is overridden (which is very likely for
custom gateway implementations).

This commit checks if method_type has been overridden, and if so, uses
that instead of the default value of partial_name.