Add payment cancellation class
This new class handles cancellations of payments called by order cancel.

This new class changes the way we cancel payments in that it asks the payment
method to try a void. If that succeeds it handles the void as always. But if
that fails - because the transaction is in a un-voidable state - it creates
a refund about the amount instead.