Remove admin's ability to edit countries and states
Allow admins to edit this part of the system could have dramatic
consequences on a store. Developers will still be able to edit these
resources through a migration or directly in the console for extreme
edge cases.

This may also be a step in the direction of using an externally sourced
service such as Carmen.