Preload assets before backend specs
We've been seeing some builds fail with "Request failed to reach server"
on the first request.

This happens because we have quite a few assets and it can take some
time to compile them. Since this was done on the first request,
this sometimes caused poltergeist to error.

This commit gets sprockets-rails to preload the assets before the suite
is run (if there is a feature spec included in the run).

This has the added advantage of not including the time to precompile
specs as part of the first spec's runtime. Running rspec --profile=10
should now be more useful in showing actually slow specs.