Remove the AJAX new variant form
This is a simplification of this admin interface in order to prepare to
switch the admin to a responsive layout.

This optimization had a weird side effect in that it fully replaced the
content on the page instead of loading the form in, say, a modal. This
would mean that hitting "Cancel" on the form would actually trigger a
page load instead of just removing the form from the page.