Use Bootstrap classnames in new admin layout
In a previous commit the `make-row()` mixin was removed from `.content`
causing the sidebar to not have enough room on the right side of the
screen. Now that we're moving to the new layout full-time we can just
add the necessary Bootstrap classes directly to the markup.

Special note: we're explicitly not using the `.container` class because
we don't want to use Bootstrap's built-in breakpoints.