Add failing spec to demonstrate issue
Some payment methods that redirect to a third-party site don't have the
concept of "capturing" a payment. It is expected that after the user
completes their checkout flow, the payment is completed. For these types
of payments, we'd prefer to complete them as soon as we return to our
Solidus site, since the provider considers them complete.

For this case, when we call `Spree::Order#add_store_credit_payments`,
the store credits will be created successfully, but there is a check at
the end that sums the payment amount of all `checkout` or `pending`
payments and adds an error if it doesn't match the order total. Since
our payment is `completed`, it doesn't get included and the totals don't
match, which incorrectly raises an error.