Better restrict deleting users with orders
Previously, there was a before_destroy on LegacyUser which would prevent
delting users which had an attached order by raiseing
DestroyWithOrdersError.

LegacyUser shouldn't be used in real stores, so this is pointless.

This replaces that check with dependent: :restrict_with_exception inside
of UserMethods, which will restrict deletions of any user model (like
Spree::User from solidus_auth_devise). This also updates the API and
admin (the only places which handled this) to understand
DeleteRestrictionError instead, and to return a better error response.