Remove unneeded failing regression test
This test wasn't testing anything useful because any exception would be
swallowed by error_during_processing.

Inevitably, this test was failing because an order in a state which
doesn't exist is invalid and state_machines was throwing an error on
`can_complete?`

This regression test existed because previously some code needed to
check if there was a template available for the specific state the order
was in. That behaviour was removed in
3af26e2c0f9aa9613a49cd9557ca150ba6514373, so we no longer need this
specific regression test.