Remove Api::BaseController#error_during_processing
Previously Api::BaseController would swallow all exceptions (or all
StandardError as of Solidus 1.0) and silently render an exception using
error_during_processing.

This has long hid errors in specs (see the last few commits for
examples) as well as in production usage.

This removes the rescue_from, the error_during_processing method, and
the configurable error_notifier. Instead, errors bubble up and are
handled by Rails middleware, as they should be.

This allows exceptions to be seen in API specs, stack traces to be seen
in development, and errors to be reported to error collection services
in production.