Create singleton class `Spree::Config`
At the moment `Spree::Config` is a constant that is being assigned as an
instance of `Spree::Core::Environment.preferences`. Using a singleton
object here is a more reasonable approach.

`Spree::Config.instance` is the actual `Spree::AppConfiguration`. Also
note that `Spree::Config#method_missing` forwards calls to
`Spree::Config.instance` so `Spree::Config` can continue to be used the
same way as before.