Exclude Bootstrap buttons from our button styling
Unfortunately we cannot currently use the Bootstrap classes on button or
input elements. This change allows us to do that by adding another
`:not()` selector to our custom button styles.

We want to keep the `@extend` so that our extensions and other legacy
content continue to have appropriately styled buttons.

Unfortunatley we cannot simply add `:not(.btn)` because we then
`@extend .btn` and Sass very quickly gets confused as to what we're
asking it to do. Therefor we need to use `:not([class*="btn"])` to
essentially trick Sass into ignoring our rather unique use case.