Assign order on package and packer on initialize
When we create and split packages, we know upfront which order
these packages belong to. However, prior to this commit we did not
assign the known order to our stock package and stock packer, which
made it necessary to probe the packages content items (if any were present)
for the order associated with the package.

This commit changes the signature of the stock package and the stock
packer such that both take an order as first, mandatory argument.

Adds some YARD docs to the packer initialization method.