Fix hound comment on defining `respond_to_missing?`
When defining `method_missing`, `respond_to_missing?` should be defined.
If it is not defined we can get some unexpected behaviour. The
`Spree::Config#method` call could give us a `NameError` exception for
methods that seem to be defined on `Spree::Config` (ones that are sent
to `Spree::Config.instance` via `#method_missing`).

For example, `Spree::Config.searcher_class` will work, but
`Spree::Config.method(:searcher_class)` will give a `NameError`.

A thoughtbot article on this:
https://robots.thoughtbot.com/always-define-respond-to-missing-when-overriding