Fix setting checkout steps context applying coupon codes
Ref: #2155

In case of errors applying coupon codes, the edit view is
rendered and controller callbacks execution ends.

But checkout steps specific ivars are set with `setup_for_current_state`
callbacks which happens after coupon code application.

This commit forces calling that method if we are stopping
callbacks execution after a coupon application error.

Since apply_coupon_code is also called from `OrdersController`
a respond_to? is required to be sure we only call it when needed.