Use product image template on frontend
Switches all product images to render the template as opposed to an
image_tag. This way the frontend views are less tied to specifics of the
models that they are rendering images for. For example, frontend views
no longer need to know that a model will have an attribute `attachment`.