Extract Packer Splitter building to SplitterChain
Previously the Stock::Packer was responsible for building a chain of
Splitter objects to split objects after packaging. This required some
precise behaviour (like order of initialization) as well as edgecase
handling (empty array of splitters) which wasn't well tested and should
be outside the responsibilities of the Packer itself.

This extracts the behaviour to Spree::Stock::SplitterChain, which
handles building and calling the splitters.