Deprecate passing Packer to Splitters
Previously, Splitters were tightly coupled to Stock::Packer without good
reason. None of the Splitters in core use this, and I don't expect it to
be used or needed in custom splitters.

This commit replaces the first argument of Splitter::Base#initialize
with StockLocation. Passing a Packer will work, but print a deprecation
warning.