Create Stock::Availability class
This commit adds Spree::Stock::Availability, a class to manage checking
stock availability efficiently for a set of Variants and StockLocations.

This is similar to the existing Spree::Stock::Quantifier, but more
efficient by checking multiple variants at once. This needs only two
queries to check stock for any number of variants and stock_locations
(though passing fewer will have faster queries).

This would be a good candidate to be a pluggable class. It could in
theory be overridden to provide oft-requested features like
stock reservations and querying other sources (like 3PL or ERP). I
haven't done this at this time because the interface isn't the best
suited to be generic and instead is just the most convenient for the
one place its used now. To do this the class would also have to include
unstocking and restocking.